Changed calls to backend to fix pie chart
Let them eat pie. I know that there is still some stuff to do with the
heatmap, changed to the right topic but we still need to implement an
API endpoint for "heatmap" and also change the javascript to be able to
iterate over the return and set up the map correctly.